now use MiniCPM_o_2_6Tokenizer and MiniCPM_o_2_6TokenizerFast; move MiniCPMVImageProcessor to image_processing_minicpm; remove token parameters in MiniCPMVImageProcessor
