+++ src/transformers/models/minicpm_o_2_6/tokenization_minicpm_o_2_6.py
@@ -13,21 +13,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import TYPE_CHECKING
+from transformers import Qwen2Tokenizer
 
-from ...utils import _LazyModule
-from ...utils.import_utils import define_import_structure
+class MiniCPM_o_2_6Tokenizer(Qwen2Tokenizer):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
 
 
-if TYPE_CHECKING:
-    from .configuration_minicpm_o_2_6 import *
-    from .image_processing_minicpm import *
-    from .modeling_minicpm_o_2_6 import *
-    from .processing_minicpm_o_2_6 import *
-    from .tokenization_minicpm_o_2_6_fast import *
-    from .tokenization_minicpm_o_2_6 import *
-else:
-    import sys
-
-    _file = globals()["__file__"]
-    sys.modules[__name__] = _LazyModule(__name__, _file, define_import_structure(_file), module_spec=__spec__)
+__all__ = ["MiniCPM_o_2_6Tokenizer"]
